requirements.txt
pillow~=12.0.0
pystray~=0.19.5
Telethon~=1.42.0
python-dotenv~=1.2.1